+++ src/ramka.h
@@ -15,7 +15,7 @@ typedef struct GOC_StFrame
 	GOC_STRUCT_FRAME;
 } GOC_StFrame;
 
-int goc_frameListener(GOC_HANDLER, GOC_MSG, void*, unsigned int);
+int goc_frameListener(GOC_HANDLER, GOC_MSG, void*, uintptr_t);
 int goc_frameSetPoints(GOC_HANDLER u, char *p);
 int goc_frameDrawer(GOC_StElement *ramka, const char *znak);
 GOC_FUN_FRAMEDRAWER *goc_elementGetFrameDrawer(GOC_HANDLER uchwyt);
